requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
aiosqlite==0.21.0
alphashape==1.3.1
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
Brotli==1.1.0
certifi==2025.7.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
click-log==0.4.0
Crawl4AI==0.7.1
cryptography==45.0.5
cssselect==1.3.0
curl_cffi==0.12.0
deprecation==2.1.0
distro==1.9.0
docstring_parser==0.16
duckduckgo_search==8.1.1
fake-http-header==0.3.5
fake-useragent==2.2.0
feedparser==6.0.11
filelock==3.18.0
frozendict==2.4.6
frozenlist==1.7.0
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.3
h11==0.16.0
h2==4.2.0
hf-xet==1.1.5
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.4
humanize==4.12.3
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
lancedb==0.24.1
lark==1.2.2
litellm==1.74.4
lxml==5.4.0
lxml_html_clean==0.4.2
markdown-it-py==3.0.0
markdown2==2.5.3
markdownify==1.1.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.3
multitasking==0.0.11
networkx==3.5
newspaper4k==0.9.3.1
nltk==3.9.1
numpy==2.3.1
openai==1.97.0
overrides==7.7.0
packaging==25.0
pandas==2.3.1
peewee==3.18.2
pgvector==0.4.1
phidata==2.7.10
pillow==11.3.0
platformdirs==4.3.8
playwright==1.53.0
primp==0.15.0
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
psycopg2-binary==2.9.10
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyee==13.0.0
Pygments==2.19.2
pyOpenSSL==25.1.0
pypdf==5.8.0
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
rank-bm25==0.2.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-file==2.1.0
rich==14.0.0
rpds-py==0.26.0
rtree==1.4.0
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
sentence-transformers==5.0.0
sgmllib3k==1.0.0
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.7
SQLAlchemy==2.0.41
sympy==1.14.0
tantivy==0.24.0
tf-playwright-stealth==1.2.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tldextract==5.3.0
tokenizers==0.21.2
tomli==2.2.1
torch==2.2.2
tqdm==4.67.1
transformers==4.53.2
trimesh==4.7.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
xxhash==3.5.0
yarl==1.20.1
yfinance==0.2.65
zipp==3.23.0
